+++ Link Scrapper/Readme.md
@@ -1,0 +1,17 @@
+# LINK SCRAPPER
+
+
+- It is used to scrape links from any website and display it.
+
+
+## Setup instructions
+
+Any PC with python 3 installed can run this code.
+
+
+## Output
+image.png
+
+## Author(s)
+
+Tanya Mohanka
